Implement Command Driven Method in Selenium tests
In this commit, a Command Driven Method was added to the TestPage function in SeleniumTestsPOM.cs. This new method submits the form using the button's Id. Simultaneously, some changes were made in the BaseCommand.cs file where the parameters' names were changed for better understanding. Additionally, the "Submit" button in the index.html file was given an Id to aid the Command Driven Method to locate it.